functions/PolicyMonitorComplianceState/tests/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity==1.10.0
azure-mgmt-policyinsights==1.0.0
black==22.6.0
parameterized==0.8.1
pytest==7.1.2
requests==2.26.0
